+++ zfit/core/space.py
@@ -16,6 +16,7 @@ import tensorflow as tf
 from tensorflow.python.util.deprecation import deprecated
 
 import zfit
+import zfit.z.numpy as znp
 
 from .. import z
 from ..settings import ztypes
@@ -159,8 +160,8 @@ def inside_rect_limits(x, rect_limits):
     lower, upper = z.unstack_x(rect_limits, axis=0)
     lower = z.convert_to_tensor(lower)
     upper = z.convert_to_tensor(upper)
-    below_upper = tf.reduce_all(input_tensor=tf.less_equal(x, upper), axis=-1)  # if all obs inside
-    above_lower = tf.reduce_all(input_tensor=tf.greater_equal(x, lower), axis=-1)
+    below_upper = znp.all(tf.less_equal(x, upper), axis=-1)  # if all obs inside
+    above_lower = znp.all(tf.greater_equal(x, lower), axis=-1)
     inside = tf.logical_and(above_lower, below_upper)
     return inside
 
@@ -188,7 +189,7 @@ def _sanitize_x_input(x, n_obs):
         if x.shape.ndims == 0:
             x = tf.broadcast_to(x, (1, 1))
         else:
-            x = tf.expand_dims(x, axis=-1)
+            x = znp.expand_dims(x, axis=-1)
     if tf.get_static_value(x.shape[-1]) != n_obs:
         raise ShapeIncompatibleError("n_obs and the last dim of x do not agree. Assuming x has shape (..., n_obs)")
     return x
@@ -1853,7 +1854,7 @@ class Space(BaseSpace):
             x_sub = self.reorder_x(x, **reorder_kwargs)
             x_inside = limit.inside(x_sub)
             xs_inside.append(x_inside)
-        all_inside = tf.reduce_all(xs_inside, axis=0)
+        all_inside = znp.all(xs_inside, axis=0)
         return all_inside
 
     @property  # TODO(discussion): depreceate 1d limits? or keep?
@@ -2056,8 +2057,8 @@ def combine_spaces(*spaces: Iterable[Space]):
                     unique_coords.remove(coord)
             elif coord in non_unique_coords:
                 non_unique_spaces = [space for space in spaces if coord in get_coord(space, using_obs)]
-                common_coords_non_unique = list(set.intersection(*[set(get_coord(space, using_obs))
-                                                                   for space in non_unique_spaces]))
+                common_coords_non_unique = list(set.intersection(*(set(get_coord(space, using_obs))
+                                                                   for space in non_unique_spaces)))
                 # do the below to check if we can take the subspace
                 non_unique_subspaces = [space.get_subspace(obs=common_coords_non_unique if using_obs else None,
                                                            axes=None if using_obs else common_coords_non_unique)
@@ -2665,7 +2666,7 @@ class MultiSpace(BaseSpace):
 
     def _inside(self, x, guarantee_limits):
         inside_limits = [space.inside(x, guarantee_limits=guarantee_limits) for space in self]
-        inside = tf.reduce_any(input_tensor=inside_limits, axis=0)  # has to be inside one limit
+        inside = znp.any(inside_limits, axis=0)  # has to be inside one limit
         return inside
 
     def __iter__(self) -> ZfitSpace: